refactor(report): split composer into context loader merger assembler persistence
